+++ soa/grayscale.cpp
@@ -8,20 +8,20 @@
 using namespace std;
 
 struct PixelNormalized {
-    vector<vector<float>> blue;
-    vector<vector<float>> green;
-    vector<vector<float>> red;
-    vector<vector<float>> gray;
+    vector<vector<double>> blue;
+    vector<vector<double>> green;
+    vector<vector<double>> red;
+    vector<vector<double>> gray;
 };
 
 //This function is used to normalize the pixel values
 //from 0 to 255 to 0.0 to 1.0
-float Normalize( uint8_t& mypixel) {
+double Normalize( uint8_t& mypixel) {
     return mypixel / 255.0;
 }
 
 //This function is used to correct the linear intensity of the pixel
-float Intensidad_lineal ( float& mypixel ) {
+double Intensidad_lineal ( double& mypixel ) {
     if (mypixel  <= 0.04045) {
         return mypixel / 12.92;
     } else {
@@ -30,12 +30,12 @@ float Intensidad_lineal ( float& mypixel ) {
 }
 
 //This function converges the 3 values of the pixel to a single value(gray)
-float Transformacion_lineal( float& mypixel_red, float& mypixel_green, float& mypixel_blue) {
+double Transformacion_lineal( double& mypixel_red, double& mypixel_green, double& mypixel_blue) {
     return 0.2126 * mypixel_red + 0.7152 * mypixel_green + 0.0722 * mypixel_blue;
 }
 
 //This function is used to apply the gamma correction of the pixel
-float correccion_gamma ( float& mypixel ) {
+double correccion_gamma ( double& mypixel ) {
     if (mypixel  <= 0.0031308) {
         return mypixel * 12.92;
     }
@@ -45,7 +45,7 @@ float correccion_gamma ( float& mypixel ) {
 }
 
 //This function is used to denormalize the pixel value
-uint8_t Denormalize(float& mypixel_gray) {
+uint8_t Denormalize(double& mypixel_gray) {
     return mypixel_gray * 255;
 }
 
@@ -53,10 +53,10 @@ int ImageSOA::ToGrayScale() {
     auto start = std::chrono::high_resolution_clock::now();
     
     PixelNormalized bmpPixelsDataNormalized;
-    bmpPixelsDataNormalized.blue.resize(bmpPixelsData.blue.size(), vector<float>(bmpPixelsData.blue[0].size()));
-    bmpPixelsDataNormalized.green.resize(bmpPixelsData.green.size(), vector<float>(bmpPixelsData.green[0].size()));
-    bmpPixelsDataNormalized.red.resize(bmpPixelsData.red.size(), vector<float>(bmpPixelsData.red[0].size()));
-    bmpPixelsDataNormalized.gray.resize(bmpPixelsData.blue.size(), vector<float>(bmpPixelsData.blue[0].size()));
+    bmpPixelsDataNormalized.blue.resize(bmpPixelsData.blue.size(), vector<double>(bmpPixelsData.blue[0].size()));
+    bmpPixelsDataNormalized.green.resize(bmpPixelsData.green.size(), vector<double>(bmpPixelsData.green[0].size()));
+    bmpPixelsDataNormalized.red.resize(bmpPixelsData.red.size(), vector<double>(bmpPixelsData.red[0].size()));
+    bmpPixelsDataNormalized.gray.resize(bmpPixelsData.blue.size(), vector<double>(bmpPixelsData.blue[0].size()));
     
     for (long unsigned int i = 0; i < bmpPixelsData.blue.size(); i++) { //Read all pixels and convert them to grayscale
         for (long unsigned int j = 0; j < bmpPixelsData.blue[0].size(); j++) {
